+++ pkg/tarutil/tarutil.go
@@ -1,0 +1,43 @@
+/*
+Copyright © 2021 VMware
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+package tarutil
+
+import (
+	"archive/tar"
+	"bytes"
+	"io"
+	"strings"
+)
+
+func ExtractFilesFromTarball(filenames map[string]string, tarf *tar.Reader) (map[string]string, error) {
+	ret := make(map[string]string)
+	for {
+		header, err := tarf.Next()
+		if err == io.EOF {
+			break
+		}
+		if err != nil {
+			return ret, err
+		}
+
+		for id, f := range filenames {
+			if strings.EqualFold(header.Name, f) {
+				var b bytes.Buffer
+				io.Copy(&b, tarf)
+				ret[id] = b.String()
+				break
+			}
+		}
+	}
+	return ret, nil
+}
